chore: scrub residual hostname/hardware refs + untrack CI artifacts
Replace all `game-client-1` hostname mentions in comments/docs/CI with
generic wording ("the deploy host", "Arch/CachyOS", "CachyOS / glibc 2.43").
No functional config changed — comments and documentation only.

Remove committed CI cruft (aqtinstall.log + tmp/shots/lane-c-fix/*.png)
from the working tree and index. Add `tmp/` and `target/` to .gitignore.

Harden lint.yml auto-format push step: unstage everything before the QML
`git add` to prevent accidental inclusion of stale working-tree files.

Fix pre-existing cross-platform type mismatch in daemon/src/system.rs
(statvfs field types are u32 on Linux, u64 on macOS) and gate
MAX_DOWNLOAD_BYTES on `cfg(target_os = "linux")` to clear a dead-code
warning on macOS. All 201 daemon tests pass; cargo clippy -D warnings clean.

Note: history purge of tmp/ + aqtinstall.log blobs is a follow-up step
(git filter-repo) to be run after this PR merges.

+++ daemon/src/system.rs
@@ -201,10 +201,16 @@ fn statvfs_entry(mountpoint: &str) -> Option<MountEntry> {
     // SAFETY: statvfs succeeded, so stat is fully initialised.
     let stat = unsafe { stat.assume_init() };
 
-    let bsize = stat.f_bsize;
-    let size = stat.f_blocks * bsize;
-    let avail = stat.f_bavail * bsize;
-    let used = size.saturating_sub(stat.f_bfree * bsize);
+    // f_bsize/f_blocks/f_bavail/f_bfree are u32 on Linux and u64 on macOS;
+    // cast uniformly to u64 so arithmetic compiles on both platforms.
+    #[allow(clippy::unnecessary_cast)]
+    let bsize = stat.f_bsize as u64;
+    #[allow(clippy::unnecessary_cast)]
+    let size = stat.f_blocks as u64 * bsize;
+    #[allow(clippy::unnecessary_cast)]
+    let avail = stat.f_bavail as u64 * bsize;
+    #[allow(clippy::unnecessary_cast)]
+    let used = size.saturating_sub(stat.f_bfree as u64 * bsize);
     let pct = if size > 0 {
         ((used as f64 / size as f64) * 100.0).round() as u8
     } else {
